Refactoring into options and utils
